requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
web3==6.12.0
python-dotenv==1.0.0
Werkzeug==2.3.7
py-solc-x==1.1.1